+++ src/sd/sd_client.cpp
@@ -156,6 +156,7 @@ public:
                                      RequestId(SOMEIP_SD_CLIENT_ID, next_multicast_session_id()),
                                      MessageType::NOTIFICATION,
                                      ReturnCode::E_OK);
+        someip_message.set_interface_version(SOMEIP_SD_INTERFACE_VERSION);
         auto serialized = sd_message.serialize();
         if (serialized.empty()) {
             platform::ScopedLock const lock(pending_finds_mutex_);
@@ -254,6 +255,7 @@ public:
                                      RequestId(SOMEIP_SD_CLIENT_ID, next_multicast_session_id()),
                                      MessageType::NOTIFICATION,
                                      ReturnCode::E_OK);
+        someip_message.set_interface_version(SOMEIP_SD_INTERFACE_VERSION);
         someip_message.set_payload(std::move(serialized));
 
         transport::Endpoint const multicast_endpoint(config_.multicast_address, config_.multicast_port);
@@ -291,6 +293,7 @@ public:
                                      RequestId(SOMEIP_SD_CLIENT_ID, next_multicast_session_id()),
                                      MessageType::NOTIFICATION,
                                      ReturnCode::E_OK);
+        someip_message.set_interface_version(SOMEIP_SD_INTERFACE_VERSION);
         someip_message.set_payload(std::move(serialized));
 
         transport::Endpoint const multicast_endpoint(config_.multicast_address, config_.multicast_port);